Update Edit.js w/ validation calibration
With the errors list object set in state, the catch for the POST request now accrues a list of errors, then passes them into the form component to show the user before the form can submit. Last change to <path> element is html-jsx property translation.
TL;DR: same as the Create.js changes in previous commit